// src/types/bookmarks.d.ts
export interface BookmarkData {
  title: string;
  creator: string;
  url: string;
  tags: string[];
}

export interface ArticleData {
  title: string;
  author: string;
  site: string;
  url: string;
  tags: string[];
}

export type PageData = ArticleData | BookmarkData;
// ...
export interface RecordData {
  id: string;
  author?: string;
  answer?: string;
  title?: string;
  tweet?: string;
  content?: string;
  creator?: string;
  question?: string;
  site?: string;
  dead?: boolean;
  subreddit?: string;
  tags: string[];
  url: string;
  user?: string;
  archive?: string;
}

export interface RecordColumnAggregateCount {
  [key: string]: number;
}

export type CountColumn =
  | 'author'
  | 'creator'
  | 'site'
  | 'dead'
  | 'subreddit'
  | 'tags'
  | 'user';

export type Tables =
  | 'articles'
  | 'comics'
  | 'podcasts'
  | 'reddits'
  | 'tweets'
  | 'videos';

export interface BookmarkingResponse {
  success: boolean;
  message: string;
  source: string;
}